ByName and regex endpoints, Populate functionality

+++ web/src/routes/package.js
@@ -20,7 +20,6 @@ const JSZip = require('jszip')
 const PackageData = require('../models/packageData')
 const Package = require('../models/package')
 const PackageRating = require('../models/packageRating')
-// const Error = require('../models/error')
 const PackageName = require('../models/packageName')
 const PackageRegEx = require('../models/packageRegEx')
 const PackageMetadata = require('../models/packageMetadata')
@@ -34,7 +33,7 @@ const execFile = util.promisify(require('node:child_process').execFile);
 
 // for Github REST API calls to extract info about modules
 const axios = require('axios')
-//const { environment } =  require('../461_CLI/environment/environment');
+const { environment } =  require('../461_CLI/environment/environment');
 
 // Here we define the routes for this endpoint
 // Per spec, this POST: Creates package
@@ -132,6 +131,8 @@ package_router.post('/', async (req,res) => {
                         })
 
                         await newPackageHistoryEntry.save()
+
+                        newPackage = await Package.findById(newPackage._id).populate('data').populate('metadata')
 
                         res.status(201).json(newPackage)
                     }
@@ -249,6 +250,8 @@ package_router.post('/', async (req,res) => {
                             })
 
                             await newPackageHistoryEntry.save()
+
+                            newPackage = await Package.findById(newPackage._id).populate('data').populate('metadata')
     
                             res.status(201).json(newPackage)
                         }
@@ -547,24 +550,33 @@ package_router.get('/:id/rate', async(req,res) => {
 //          - 400: There is missing field(s) in the PackageName/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid.
 //          - 404: No such package.
 package_router.get('/byName/:name', async(req,res) => {
-    const Name = await PackageName.findOne({PackageName: req.params.name})
-    let isValid = true
-    // Created test PackageHistoryEntry 
-    const test = {
-        User: '1234',
-        Date: new Date(),
-        PackageMetadata: 'abcd',
-        Action: 'CREATE'
-    }
-    try {
-        await Name.validate()
-    } catch (err){
-        isValid = false
+    let md;
+
+    // Check package exists
+    if(!(md = await PackageMetadata.findOne({ Name: req.params.name }))) {
         res.status(404).json({ message: 'No such package.' })
     }
-    if (isValid){
-        res.json(Name)
-        //res.json(test)
+    else {
+        let isValid = true
+        
+        // Validate package
+        try {
+            await md.validate()
+        } catch {
+            isValid = false
+            res.status(400).json({ message: 'There is missing field(s) in the PackageName or it is formed improperly.' })
+        }
+        if (isValid){
+            let output_array = []
+
+            // Create array of packages matching name
+            try {
+                output_array = await PackageHistoryEntry.find({ PackageMetaData: md._id }).exec()
+                res.status(200).json(output_array)
+            } catch {
+                res.status(500).json({ message: 'Unexpected error.' })
+            }
+        }
     }
 })
 
@@ -577,35 +589,36 @@ package_router.get('/byName/:name', async(req,res) => {
 //          - 400: There is missing field(s) in the PackageName/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid.
 //          - 404: Package does not exist.
 package_router.delete('/byName/:name', async(req,res) => {
-    const name = await PackageName.findOne({PackageName: req.params.name})
-    //console.log(Name)
-    let isValid = true
-    try {
-        await name.validate()
-    } catch (err){
-        isValid = false
-        res.status(404).json({ message: 'Package does not exist.' })
+    let md
+
+    // Find package 
+    if(!(md = await PackageMetadata.findOne({ Name: req.params.name }))) {
+        res.status(404).json({ message: 'No such package.' })
     }
-    if (isValid){
-        res.status(200).json({ message: 'Package is deleted.' })
-        const package = await Package.findOne(await PackageMetadata.findOne(await PackageName.findOne({PackageName: req.params.name})))
-        const md = await PackageMetadata.findById(package.metadata)
-        const data = await PackageData.findById(package.data)
-        const id = await PackageID.findById(md.ID)
-        await Promise.all([
-            // Is this all I have to delete 
-            Package.deleteMany(package._id),
-            PackageData.deleteMany(data._id),
-            //PackageHistoryEntry.deleteMany(name._id),
-            PackageID.deleteMany(id._id),
-            PackageMetadata.deleteMany(md._id),
-           // PackageQuery.deleteMany(name._id),
-            //PackageRating.deleteMany(name._id),
-            //SemverRange.deleteMany(name._id),
-            //User.deleteMany(name._id),
-            PackageName.deleteMany(name._id)
-        ]);
-        
+    else {
+        let isValid = true
+
+        // Validate
+        try {
+            await md.validate()
+        } catch {
+            isValid = false
+            res.status(400).json({ message: 'There is missing field(s) in the Name or it is formed improperly.' })
+        }
+        if (isValid){
+            // Find Package and PackageData documents
+            const package = await Package.findOne({ metadata: md._id })
+            const data = await PackageData.findById({ _id: package.data })
+            
+            // Delete related packages
+            await Promise.all([
+                Package.deleteMany(package._id),
+                PackageData.deleteMany(data._id),
+                PackageHistoryEntry.deleteMany({ PackageMetadata: md._id }),
+                PackageMetadata.deleteMany(md._id),
+            ]);
+            res.status(200).json({ message: 'Package is deleted.' })
+        }
     }
 })
 
@@ -618,29 +631,32 @@ package_router.delete('/byName/:name', async(req,res) => {
 //          - 400: There is missing field(s) in the PackageRegEx/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid.
 //          - 404: No package found under this regex.
 package_router.post('/byRegEx', async(req,res) => {
-    const newPackageDataSchema = new PackageData(req.body)
+    // Load reg ex into schema
+    const newPackageRegEx = await new PackageRegEx(req.body)
+
+    // Validate
+    try {
+        await newPackageRegEx.validate()
+    } catch {
+        res.status(400).json({ message: 'There is missing field(s) in the PackageRegEx or it is formed improperly.' })
+    }
+
+    // Search by reg ex
+    let output_array = []
+    try {
+        output_array = await PackageMetadata.find({ Name: { $regex: newPackageRegEx.PackageRegEx } }).exec()
+    } catch {
+        res.status(500).json({ message: 'Unknown error.' })
+    }
+
+    // Check if nothing was found
+    if(output_array.length == 0) {
+        res.status(404).json({ message: 'No package found under this regex.' })
+    }
+    else {
+        res.status(200).json(output_array)
+    }
 })
-
-// If functions are needed, name them according to operationId in spec
-// If functions needed for:
-//      - Creating package
-//          Name it: PackageCreate
-//      - Retrieving Package
-//          Name it: PackageRetrieve
-//      - Updating package
-//          Name it: PackageUpdate
-//      - Deleting package
-//          Name it: PackageDelete
-//      - Rating Package
-//          Name it: PackageRate
-//      - Getting package by name
-//          Name it: PackageByNameGet
-//      - Deleting package by name
-//          Name it: PackageByNameDelete
-//      - Getting package by RegEx
-//          Name it: PackageByRegExGet
-// Replace these comments related to the function with comment describing function
-
 
 // Export the router as a module so other files can use it
 module.exports = package_router
